requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# telegram-top-posts-bot/main.py: 7
chatbase == 0.2.1

# telegram-top-posts-bot/utils.py: 1,2
instaloader == 4.2.9

# telegram-top-posts-bot/utils.py: 3
pandas == 0.24.2

# telegram-top-posts-bot/mailing_users.py: 4
# telegram-top-posts-bot/main.py: 8
python_telegram_bot == 12.0.0b1

# telegram-top-posts-bot/utils.py: 5
tabulate == 0.8.4

# telegram-top-posts-bot/mailing_users.py: 6
# telegram-top-posts-bot/main.py: 14
tinydb == 3.14.1

# telegram-top-posts-bot/mailing_users.py: 7
tqdm == 4.32.1
